+++ src/main/java/es/in2/issuer/domain/model/enums/CredentialStatus.java
@@ -1,8 +1,8 @@
 package es.in2.issuer.domain.model.enums;
 
 public enum CredentialStatus {
+    WITHDRAWN, //fixme: Deprecated. Old status. Necessary to maintain retro compatibility.
     DRAFT,
-    WITHDRAWN,
     ISSUED,
     PEND_DOWNLOAD,
     VALID,