+++ src/main/java/es/in2/issuer/domain/service/CredentialOfferService.java
@@ -1,10 +1,10 @@
 package es.in2.issuer.domain.service;
 
-import es.in2.issuer.domain.model.dto.CustomCredentialOffer;
+import es.in2.issuer.domain.model.dto.CredentialOfferData;
 import es.in2.issuer.domain.model.dto.Grant;
 import reactor.core.publisher.Mono;
 
 public interface CredentialOfferService {
-    Mono<CustomCredentialOffer> buildCustomCredentialOffer(String credentialType, Grant grant);
+    Mono<CredentialOfferData> buildCustomCredentialOffer(String credentialType, Grant grant, String employeeEmail, String pin);
     Mono<String> createCredentialOfferUri(String nonce);
 }