+++ src/test/java/es/in2/issuer/infrastructure/controller/GlobalExceptionHandlerTest.java
@@ -2,7 +2,6 @@ package es.in2.issuer.infrastructure.controller;
 
 import es.in2.issuer.domain.exception.*;
 import es.in2.issuer.domain.model.dto.CredentialErrorResponse;
-import es.in2.issuer.domain.model.dto.GlobalErrorMessage;
 import es.in2.issuer.domain.util.CredentialResponseErrorCodes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,7 +13,6 @@ import reactor.test.StepVerifier;
 
 import javax.naming.OperationNotSupportedException;
 import java.text.ParseException;
-import java.time.LocalDateTime;
 import java.util.NoSuchElementException;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -55,9 +53,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleNoSuchElementException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -199,26 +195,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleParseException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    void handleException_ReturnsInternalServerError() {
-        Exception exception = new Exception("General error occurred");
-        Mono<ResponseEntity<GlobalErrorMessage>> result = globalExceptionHandler.handleException(exception, mockWebRequest);
-        StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                    assertNotNull(responseEntity.getBody());
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR.value(), responseEntity.getBody().status());
-                    assertEquals("General error occurred", responseEntity.getBody().message());
-                    assertTrue(LocalDateTime.now().minusMinutes(1).isBefore(responseEntity.getBody().timestamp()) &&
-                                    LocalDateTime.now().plusMinutes(1).isAfter(responseEntity.getBody().timestamp()),
-                            "Timestamp is not within the expected range");
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -229,9 +206,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleBase45Exception(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -242,9 +217,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleCreateDateException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -255,9 +228,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleSignedDataParsingException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -268,9 +239,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleSignedDataParsingException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -281,9 +250,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleParseCredentialJsonException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -294,9 +261,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleTemplateReadException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -307,9 +272,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleProofValidationException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -320,9 +283,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleNoCredentialFoundException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -333,9 +294,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handlePreAuthorizationCodeGetException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -346,9 +305,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleCustomCredentialOfferNotFoundException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode()))
                 .verifyComplete();
     }
 
@@ -359,9 +316,7 @@ class GlobalExceptionHandlerTest {
         Mono<ResponseEntity<Void>> result = globalExceptionHandler.handleNonceValidationException(exception);
 
         StepVerifier.create(result)
-                .assertNext(responseEntity -> {
-                    assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-                })
+                .assertNext(responseEntity -> assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode()))
                 .verifyComplete();
     }